Initial code dump
Based on my previous works in the Arduino environment the initial code
was forged. The basic idea is to have 5 channels for fan control. Three
channels are for 2 or 3 wire fans, and two channels are for 4 wire
fans. To handle the ~25 kHz requirement, Timer1 speed is changed to ~31
kHz. For the 4 wire fans a tachometer will be implemented using the
interrupt pins. Pins were also assigned for an optional LCD display.
The temperature source can be a DS18B20 sensor or an LM35, whichever is
preferred. The same SerialEvent handler will be used as in my other
projects. The if/else structure will be ugly, but it is needed for the
external control and munin integration.

Itemized:

- Create header.
- Assign pins.
- Reserve pins for LCD.
- Design serial commands.
- Setup function.
- PWM Frequency change to 31 kHz.
- Read LM35 Temperature sensor function.
- SerialEvent for external control.
- Basic for loop.

It compiles in old (1.8.19) and new (2.2.1) Arduino IDE. Arduino Nano.

Sketch uses 4618 bytes (15%) of program storage space. Maximum is 30720 bytes.
Global variables use 329 bytes (16%) of dynamic memory, leaving 1719 bytes
for local variables. Maximum is 2048 bytes.