Batch of Arduino updates
I continued the work on the code and on the circuit on the prototype board. Some functions were
removed or updated. The current idea is to use more arrays and access them by index. At the
current stage I do not want to use advanced techniques, as we have enough space and memory.
Basically the serial communication part is complete, only the munin part is left.

Itemized:

* General configuration string and bool for enabling and disabling a fan channel.
* The fan pwm pins are now defined in a single array.
* The settings are defined in a 5*8 2D array:
  * manual [0,1]
  * current pwm value [0-255]
  * pwm values [LOW[0-255], HIGH[0-255], CRITICAL[255]]
  * temp values [tLOW[0-85], tHIGH[0-85], tCRITICAL[85]]
* start interrupt work.
* set ADC reference to 1100 mV, this increases the resolution of the LM35* sensor.
* temperature source wrapper.
* all serial commands.
* automode: read temperature and update fan pwm speeds accordingly.

NOTE: I used the map() and constrain() functions to convert or limit ranges.

It compiles in old (1.8.19) and new (2.2.1) Arduino IDE. Arduino Nano.

Sketch uses 10460 bytes (34%) of program storage space. Maximum is 30720 bytes.
Global variables use 504 bytes (24%) of dynamic memory, leaving 1544 bytes for local variables. Maximum is 2048 bytes.